move config under frontend
